Cleaned up all warnings and tidied up some modules.
